requirements.txt
colomoto-jupyter
pandas
networkx
pydot
matplotlib